add tooltip component to allow override, move chat uploaded file components to subdir
