+++ frontend/lib/src/components/shared/Tooltip/Tooltip.tsx
@@ -17,7 +17,12 @@
 import React, { ReactElement, ReactNode, useRef } from "react"
 
 import { useTheme } from "@emotion/react"
-import { ACCESSIBILITY_TYPE, PLACEMENT, StatefulTooltip } from "baseui/tooltip"
+import {
+  ACCESSIBILITY_TYPE,
+  PLACEMENT,
+  PopoverOverrides,
+  StatefulTooltip,
+} from "baseui/tooltip"
 
 import {
   EmotionTheme,
@@ -49,6 +54,7 @@ export interface TooltipProps {
   inline?: boolean
   style?: React.CSSProperties
   onMouseEnterDelay?: number
+  overrides?: PopoverOverrides
 }
 
 function Tooltip({
@@ -58,6 +64,7 @@ function Tooltip({
   inline,
   style,
   onMouseEnterDelay,
+  overrides,
 }: TooltipProps): ReactElement {
   const theme: EmotionTheme = useTheme()
   const { colors, fontSizes, radii, fontWeights } = theme
@@ -119,42 +126,45 @@ function Tooltip({
       popoverMargin={10}
       onMouseEnterDelay={onMouseEnterDelay}
       overrides={{
-        Body: {
-          style: {
-            // This is annoying, but a bunch of warnings get logged when the
-            // shorthand version `borderRadius` is used here since the long
-            // names are used by BaseWeb and mixing the two is apparently
-            // bad :(
-            borderTopLeftRadius: radii.default,
-            borderTopRightRadius: radii.default,
-            borderBottomLeftRadius: radii.default,
-            borderBottomRightRadius: radii.default,
+        ...{
+          Body: {
+            style: {
+              // This is annoying, but a bunch of warnings get logged when the
+              // shorthand version `borderRadius` is used here since the long
+              // names are used by BaseWeb and mixing the two is apparently
+              // bad :(
+              borderTopLeftRadius: radii.default,
+              borderTopRightRadius: radii.default,
+              borderBottomLeftRadius: radii.default,
+              borderBottomRightRadius: radii.default,
 
-            paddingTop: "0 !important",
-            paddingBottom: "0 !important",
-            paddingLeft: "0 !important",
-            paddingRight: "0 !important",
+              paddingTop: "0 !important",
+              paddingBottom: "0 !important",
+              paddingLeft: "0 !important",
+              paddingRight: "0 !important",
 
-            backgroundColor: "transparent",
+              backgroundColor: "transparent",
+            },
+          },
+          Inner: {
+            style: {
+              backgroundColor: hasLightBackgroundColor(theme)
+                ? colors.bgColor
+                : colors.secondaryBg,
+              color: colors.bodyText,
+              fontSize: fontSizes.sm,
+              fontWeight: fontWeights.normal,
+
+              // See the long comment about `borderRadius`. The same applies here
+              // to `padding`.
+              paddingTop: "0 !important",
+              paddingBottom: "0 !important",
+              paddingLeft: "0 !important",
+              paddingRight: "0 !important",
+            },
           },
         },
-        Inner: {
-          style: {
-            backgroundColor: hasLightBackgroundColor(theme)
-              ? colors.bgColor
-              : colors.secondaryBg,
-            color: colors.bodyText,
-            fontSize: fontSizes.sm,
-            fontWeight: fontWeights.normal,
-
-            // See the long comment about `borderRadius`. The same applies here
-            // to `padding`.
-            paddingTop: "0 !important",
-            paddingBottom: "0 !important",
-            paddingLeft: "0 !important",
-            paddingRight: "0 !important",
-          },
-        },
+        ...overrides,
       }}
     >
       {/* BaseWeb manipulates its child, so we create a wrapper div for protection */}
